extract interface and 3 specific realization, that can stream element info to data, realizing single responsibility principle
