+++ CS/Snoop/CollectorExts/IElementStream.cs
@@ -1,0 +1,9 @@
+using System;
+
+namespace RevitLookup.Snoop.CollectorExts
+{
+    public interface IElementStream
+    {
+        void Stream(Type type);
+    }
+}
